Revert "cconvey/merge some upstream (#367)" (#368)
This reverts commit cfbcdab500e494c2dec4cba6ccfa8245ccc6120a.

+++ src/operator/operator_common.h
@@ -494,7 +494,7 @@ inline void LogUnimplementedOp(const nnvm::NodeAttrs& attrs,
 }
 
 class OpSignature {
-  std::vector<int64_t> eles;
+  std::vector<int> eles;
   uint64_t hash;
 
  public:
